Migrated to python
I have started the migration from a node.js app using express and mqtt to a
RESTful python application using Flask. This change was mostly architectural in
nature, moving away from a pub/sub model with MQTT to a REST API.

This change was mainly to minimize the latency between when a command was issued
to the client and when the change would occur. In the MQTT model, it could take
up to 2 seconds between issuing a command and observing a change. Now, it is
under a second.

The other reason for this change was to minimize the number of moving parts in
the system. Now, each client is a small HTTP server that follows a
standardized API. This will allow me to be more flexible in my integrations
with other services.