Added multiple client registrations
Now, a client can register as many times as it'd like without any negative
consequences. This will allow for occasional updating of client IP addresses in
micasa.py without negative consequences. The addition of the client name to the
response of /client necessitated an update to fauxmo.py.

I will endeavour to add updating to the fauxmo server when possible.