Added floorplan and thermometer data
To have a better understanding of how my home is heated, I added
an SVG floorplan, and a polling function to gather thermometer
data from each of my clients. Temperatures are encoded to a color
ramp between blue and red with between 55F and 75F.

I'm considering adding historical graphs, but that will require
the addition of d3. This could be an interesting learning
experience.